requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-countries==7.4.2
django-environ==0.9.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumberslite==8.13.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
